Update API for create_alert in kibana_alert
The API that the create_alert function was using is
deprecated. The function has been updated to use
the current API.

A couple small changes to the API call:

First, the endpoint has been changed to 'alerting/rule' as
described by the API.

Second, Some of the top level parameters in the body of the API call
have changed from camel case to snake case.

For example: NotifyWhen is now notify_when.

AlertTypeId changed to rule_type_id

It seems that the parameters being passed to the rule type
are still camel case, so those don't need to be changed at this time.