Move CRUD functions from kibana_alert module to module_utils/kibana
Moving CRUD functions to module_utils/kibana so that other
modules can make use of them if needed.

In this process, we are also merging the create and update functions,
as they are similar enough to be done in a single function. This
function will now be called ensure_alert(), and will take an argument
for the method to use (POST, PUT)