Handle different action types in kibana_fleet_host
Take an action from the user and manipulate the fleet data accordingly.

Available actions are add, overwrite and remove.

Add will compare the provided urls to the curren urls in kibana, and
add them if they don't exist

Remove will iterate over the current urls in kibana and omit them
if they exist in the provided urls.

Overwrite will ignore the current urls, and push the provided urls
to the kibana fleet.