Fix endless loop when activating security rules
The Operations CTRL and Endpoint Security CTRL Baseline Create job was
hanging for days when trying to activate the security rules. This is
because the logic in the while loop of the activate_security_rule()
function in kibana.py wasn't cycling pages correctly. The new logic has
been tested and will exit properly